requirements.txt
numpy==2.2.6
pillow==12.0.0
pygame==2.6.1
pyglm==2.8.3
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
